+++ SimpleStellaris/src/SpaceMapConfiguration.h
@@ -66,7 +66,7 @@ struct SpaceMapConfigurations
 	//Distances in A.U.
 	sf::Vector2f closeStarsBoundaries{ 0.1f, 10.f };
 	sf::Vector2f afarStarsBoundaries{ 15.f, 1000.f };
-	float sunDiameter{ 0.009346f };//0.009346f
+	double sunDiameter{ 0.009346 };//0.009346
 
 	//Planets configuration
 	//Habitable zone size
